MDToolbox.jl

MDToolbox.jl is a Julia package for molecular dynamics (MD) trajectories analysis and modeling of biomolecules. It consists of a collection of functions covering the following types of scientific computations:

I/O for trajectory, coordinate, and topology files used for MD simulation

New complex type for containing and operating molecular dynamics trajecory data

Flexible atom selections

Least-squares fitting of structures

Potential mean force (PMF) or free energy profile from scattered data

Statistical estimates (WHAM and MBAR methods) from biased data

Dimensional reductions (Principal Component Analysis, and others)

Other utility functions

Unique point of MDToolbox is that it is a pure Julia package; all codes are written with Julia. This make developers easy to maintain the package codes.

MDToolbox.jl

Reduction

Index

Reduction

MDToolbox.pca — Method

pca(X::AbstractMatrix; k=dimension) -> F

Perform principal component analysis (PCA). PCA captures degrees of freedom which have the largest variances. Input data X should belong to AbstractMatrix type and its columns corresponds to variables, rows are frames.

Returns a NamedTuple object F which contains the principal components in F.projection, the prncipal modes in the columns of the matrix F.mode, and the variances of principal components in F.variance.

If k=dimension is specified, the randomized SVD is used for reducing memory. This algorithm first project the data into a randomly selected k+2-dimensional space, then PCA is performed in the projected data. See the references for details. Note that if the dimension of X is larger than 5000, the randomized SVD is forcibly used with k=1000.

Example

julia> using MDToolbox, Plots, Statistics
julia> X = cumsum(rand(1000, 10))
julia> F = pca(X)
julia> plot(F.projection[:, 1], F.projection[:, 2])

References

Halko, N., Martinsson, P.-G., Shkolnisky, Y. & Tygert, M. 
An Algorithm for the Principal Component Analysis of Large Data Sets. 
SIAM J. Sci. Comput. 33, 2580–2594 (2011).

Index

MDToolbox.pca