MBAR (Multistate Bennett acceptance ratio)

MDToolbox.mbar — Method

mbar(u_kl; ftol=1e-8, iterations=10^2) -> F

Estimates the free energy differences of umbrella-windowed systems by using the Multistate Bennet Acceptance Ratio Method (MBAR). Let K be # of umbrellas or different ensembles. u_kl is a K x K Array whose elements are reduced bias-factor or potential energy of umbrella (or ensemble) simulation data k evaluated at umbrella (or ensemble) l.

Returns (dimensionless) free energies of umbrella-windows or ensembles f_k.

References

M. R. Shirts and J. D. Chodera, J. Chem. Phys. 129, 124105 (2008).