WHAM (Weighted Histogram Analysis Method)

MDToolbox.wham — Method

wham(h_km, bias_km; ftol=1e-10, maxiterations=10^5) -> F

Estimates (reduced) free energies of umbrella-windows and potential of mean force in data-bins by using the WHAM equations. K is # of umbrellas and M is # of bins. h_km is a K x M Array which is a histogram (data counts) of k-th umbrella data in m-th bin. bias_km is also a K x M Array which is a bias-factor of k-th umbrella-window evaluated at m-th bin-center.

Returns a NamedTuple object F whose members are F.f_k reduced relative free energies of umbrella-windows, and pmf_m reduced potential of mean force in data-bins under unbiased condition.

References

S. Kumar, D. Bouzida, R. H. Swendsen, P. A. Kollman, and J. M. Rosenberg, 
J. Comput. Chem. 13, 1011 (1992).